Split Buffer (glTF model) from our fast buffer concat implementation
Also concatenate those buffers straight into each other without
generating an early intermediate concatenation

This results in a change in the gltf file, but it's just the order of
two of the dictionary keys changing.
