+++ src/main/kotlin/models/glTF/Buffer.kt
@@ -1,38 +1,5 @@
 package models.glTF
 
-import com.fasterxml.jackson.annotation.JsonIgnore
-import java.nio.ByteBuffer
-
-class Buffer(filename: String) {
+class Buffer(filename: String, val byteLength: Int) {
     val uri = "$filename.bin"
-
-    fun getByteLength() = byteChunks.sumOf { it.limit() }
-
-    private val byteChunks = ArrayList<ByteBuffer>()
-
-    @JsonIgnore
-    fun getBytes(): ByteArray {
-        val buf = getByteBuffer()
-        val arr = ByteArray(buf.remaining())
-        buf.get(arr)
-        return arr
-    }
-
-    @JsonIgnore
-    fun getByteBuffer(): ByteBuffer {
-        val finalBytes = ByteBuffer.allocateDirect(getByteLength())
-        byteChunks.forEach { chunk ->
-            finalBytes.put(chunk)
-        }
-        finalBytes.flip()
-
-        // Might as well cache the fruit of our efforts
-        byteChunks.clear()
-        byteChunks.add(finalBytes.duplicate())
-        return finalBytes
-    }
-
-    fun addBytes(bytesToAdd: ByteBuffer) {
-        byteChunks.add(bytesToAdd)
-    }
 }
